move splitresult to build
